Add setting for local changes
If the commit parameter is set to "manual", then the tool does not touch the library and just uses the local copy

+++ src/frameworks/util.py
@@ -217,6 +217,7 @@ def git_clone(url: str, commit: str, name: str):
 
 def git_reset(commit: str, name: str):
     os.chdir(name)
-    run_subprocess(f'git reset --hard {commit}')
-    run_subprocess(f'git clean -f -x .')
+    if commit!="manual":
+        run_subprocess(f'git reset --hard {commit}')
+        run_subprocess(f'git clean -f -x .')
     os.chdir('../')
